@@ -1,14 +1,19 @@
-import { Response } from "@angular/http";
+import { HttpErrorResponse } from "@angular/common/http";
 
 import { Observable } from "rxjs/Observable";
 import 'rxjs/add/observable/throw';
 
 export class ErrorHandler {
 
-    static errorHandler(error: Response | any) {
+    static errorHandler(error: HttpErrorResponse | any) {
         let errorMessage: string;
 
-        errorMessage = error instanceof Response ? `Error ${error.status} when accessing URL ${error.url} - ${error.statusText}` : error.tostring();
+        if(error instanceof HttpErrorResponse) {
+            const body = error.error;
+            errorMessage = `${error.url}: ${error.status} - ${error.statusText || ''} ${body}`;
+        } else {
+            errorMessage = error.message ? error.message : error.toString();
+        }
 
         console.log(errorMessage);
         return Observable.throw(errorMessage);
