@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http } from '@angular/http';
+import { HttpClient, HttpParams } from '@angular/common/http';
 
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
@@ -13,29 +13,25 @@ import { MenuItem } from 'app/restaurant-detail/menu-item/menu-item.model';
 @Injectable()
 export class RestaurantsService {
 
-  constructor(private http: Http) { }
+  constructor(private http: HttpClient) { }
 
   getRestaurants(search?: string): Observable<Restaurant[]> {
-    return this.http.get(`${MEAT_API}/restaurants`, {params: {q: search}})
-      .map(response => response.json())
-      .catch(ErrorHandler.errorHandler);
+    let params = undefined;
+    if(search) {
+      params = new HttpParams().append('q', search);
+    }
+    return this.http.get<Restaurant[]>(`${MEAT_API}/restaurants`, {params});
   }
 
   getRestaurantById(id: string): Observable<Restaurant> {
-    return this.http.get(`${MEAT_API}/restaurants/${id}`)
-      .map(response => response.json())
-      .catch(ErrorHandler.errorHandler);
+    return this.http.get<Restaurant>(`${MEAT_API}/restaurants/${id}`);
   }
 
   getReviewsOfRestaurant(id: string): Observable<any> {
-    return this.http.get(`${MEAT_API}/restaurants/${id}/reviews`)
-      .map(response => response.json())
-      .catch(ErrorHandler.errorHandler);
+    return this.http.get(`${MEAT_API}/restaurants/${id}/reviews`);
   }
 
   getMenuOfRestaurant(id: string): Observable<MenuItem[]> {
-    return this.http.get(`${MEAT_API}/restaurants/${id}/menu`)
-      .map(response => response.json())
-      .catch(ErrorHandler.errorHandler);
+    return this.http.get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`);
   }
 }
